analysis-individual/EK/dsp-api/pyproject.toml
[project]
name = "capstone_benchmark"
version = "0.1.0"
description = ""
authors = [
    {name = "Elaine Keung",email = "ekeung@berkeley.edu"}
]

readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
fastapi = { extras = ["standard"], version = "^0.112.2" }
pandas = ">=2.2.3,<3.0.0"
statsmodels = ">=0.14.4,<0.15.0"
pyportfolioopt = ">=1.5.6,<2.0.0"
cvxpy = ">=1.6.2,<2.0.0"
seaborn = ">=0.13.2,<0.14.0"
numpy = ">=2.2.3,<3.0.0"
joblib = ">=1.4.2,<2.0.0"
scikit-learn = ">=1.6.1,<2.0.0"
openpyxl = ">=3.1.5,<4.0.0"
dill = ">=0.3.9,<0.4.0"
matplotlib = ">=3.10.1,<4.0.0"
fastapi-cache2 = {version = "0.1.9", extras = ["redis"]}
torch = "^2.6.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
ruff = "^0.6.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
